+++ R/cumulate.R
@@ -1,20 +1,76 @@
 #' Compute cumulative 'incidence'
 #'
-#' @description `r lifecycle::badge("deprecated")`
-#'
-#' `cumulate` has been deprecated to avoid encouraging erroneous regression
-#'   on cumulative counts.
+#' `cumulate` is an S3 generic to compute cumulative numbers, with methods
+#' for different types of objects:
+#' * default method is a wrapper for `cumsum`
+#' * `incidence` objects: computes cumulative incidence over time
 #'
 #' @param x An incidence object.
 #'
-#' @name cumulate-deprecated
-#' @keywords internal
+#' @examples
+#' dat <- data.frame(
+#'   dates = as.integer(c(0,1,2,2,3,5,7)),
+#'   groups = factor(c(1, 2, 3, 3, 3, 3, 1))
+#' )
+#'
+#' i <- incidence(dat, date_index = dates, groups = groups)
+#' i
+#'
+#' cumulative_i <- cumulate(i)
+#' cumulative_i
+#'
+#' @import data.table
 #' @export
+#'
+#' @rdname cumulate
 cumulate <- function(x) {
-  lifecycle::deprecate_stop(
-    when = "1.2.0",
-    what = "cumulate()",
-    details = "* See the 1.2.0 NEWS for more information"
-  )
+  UseMethod("cumulate", x)
 }
 
+#' @export
+#' @rdname cumulate
+cumulate.default <- function(x) {
+  cumsum(x)
+}
+
+
+#' @export
+#' @rdname cumulate
+cumulate.incidence_df <- function(x) {
+
+  # due to NSE notes in R CMD check
+  ..count_var <- NULL
+
+  is_cumulate <- attr(x, "cumulative")
+  if (is_cumulate) abort("x is already a cumulative incidence")
+
+  groups <- get_group_names(x)
+  count_var <- get_count_names(x)
+
+  dt <- !any(vapply(x, typeof, character(1)) == "list")
+  if (dt) {
+    out <- as.data.table(x)
+    if (!is.null(groups)) {
+      out[, (count_var) := lapply(.SD, cumsum), keyby = groups, .SDcols = count_var]
+    } else {
+      out[, (count_var) := lapply(.SD, cumsum), .SDcols = count_var]
+    }
+    setDF(out)
+  } else {
+    out <- x
+    if (!is.null(groups)) {
+      out <- grouped_df(out, groups)
+      out <- mutate(out, across(all_of(count_var), cumsum))
+      out <- ungroup(out)
+      out <- out[order(out[groups], out$date_index),]
+    } else {
+      out[count_var] <- lapply(out[count_var], cumsum)
+    }
+  }
+
+  nms <- names(out)
+  attributes(out) <- attributes(x)
+  names(out) <- nms
+  if (inherits(out, "incidence2")) attr(out, "cumulative") <- TRUE
+  out
+}