+++ R/plot.R
@@ -422,6 +422,9 @@ ylabel <- function(x, ylab) {
       }
 
     }
+    if (isTRUE(attr(x, "cumulative"))) {
+      ylab <- sub("incidence", "cumulative incidence", ylab)
+    }
     first_letter <- substring(ylab, 1, 1)
     substring(ylab, 1, 1) <- toupper(first_letter)
   }